backend/requirements.txt
asgiref==3.10.0
astroid==3.3.11
colorama==0.4.6
dill==0.4.0
Django==5.1.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
importlib_metadata==8.7.0
isort==6.1.0
mccabe==0.7.0
numpy==2.3.4
pandas==2.3.3
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
zipp==3.23.0
